pyproject.toml
[project]
name = "slack-status"
version = "0.1.0"
description = "A CLI tool to update your Slack status with emoji, text, and optional expiration"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your@email.com" }]
dependencies = [
    "typer[all]",
    "requests",
    "python-dotenv"
]

[project.scripts]
slack-status = "slack_status_updater.cli:app"

[tool.setuptools]
packages = ["slack_status_updater"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest",
    "requests-mock",
    "typer[all]"
]
